amended shadow on custom cell, moved all code into extensions for clean controllers
