+++ MyFifaTracker/Controllers/TeamsController.swift
@@ -9,27 +9,14 @@
 import UIKit
 import CoreData
 
-class TeamsController: UITableViewController, createTeamControllerDelegate {
-    
-    func editTeam(team: Teams) {
-        let row = teams.index(of: team)
-        let reloadPath = IndexPath(row: row!, section: 0)
-        tableView.reloadRows(at: [reloadPath], with: .middle)
-    }
-    
-    
-    func addTeam(team: Teams) {
-        teams.append(team)
-        let newIndexPath = IndexPath(row: teams.count - 1, section: 0)
-        tableView.insertRows(at: [newIndexPath], with: .middle)
-    }
+class TeamsController: UITableViewController {
     
     let cellId = "cellId"
     var teams = [Teams]()
     
     override func viewDidLoad() {
         super.viewDidLoad()
-
+        
         tableView.backgroundColor = .lightRedBackGround
         tableView.separatorStyle = .none
         tableView.register(TeamsCustomCell.self, forCellReuseIdentifier: cellId)
@@ -41,67 +28,8 @@ class TeamsController: UITableViewController, createTeamControllerDelegate {
         
     }
     
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return teams.count
-    }
-    
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: cellId, for: indexPath) as! TeamsCustomCell
-        let team = teams[indexPath.row]
-        cell.team = team
-        cell.selectionStyle = .none
-        return cell
-    }
-    
-    override func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 60
-    }
-
-    override func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath) -> [UITableViewRowAction]? {
-        let deleteAction = UITableViewRowAction(style: .destructive, title: "Delete") { (_, indexPath) in
-            let team = self.teams[indexPath.row]
-            print("attempting to delete team :-", team.name ?? "")
-            self.teams.remove(at: indexPath.row)
-            self.tableView.deleteRows(at: [indexPath], with: .automatic)
-            
-            let context = CoreDataManager.shared.persistantContainer.viewContext
-            context.delete(team)
-            
-            do {
-                try context.save()
-            } catch let err {
-                print("error deleting from coreData", err)
-            }
-            
-        }
-        
-        let editAction = UITableViewRowAction(style: .normal, title: "Edit", handler: editHandlerFunction)
-        
-        return [deleteAction, editAction]
-    }
-    
-    fileprivate func editHandlerFunction(action: UITableViewRowAction, indexPath: IndexPath) {
-        print("test edit")
-        let editComapanyController = CreateTeamController()
-        editComapanyController.team = teams[indexPath.row]
-        editComapanyController.delegate = self
-        let navController = CustomNavigationController(rootViewController: editComapanyController)
-        present(navController, animated: true, completion: nil)
-    }
-    
     fileprivate func fetchTeams() {
-        let context = CoreDataManager.shared.persistantContainer.viewContext
-        let fetchRequest = NSFetchRequest<Teams>(entityName: "Teams")
-        do {
-            let myTeams = try context.fetch(fetchRequest)
-            myTeams.forEach { (team) in
-                print(team.name ?? "")
-            }
-            self.teams = myTeams
-            self.tableView.reloadData()
-        } catch let err {
-            print("unable to fetch teams", err)
-        }
+        self.teams = CoreDataManager.shared.fetchTeams()
     }
     
     @objc fileprivate func handleAddTeam() {
